+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
   - #### 🚀 Enhancements:
     - Refactor main module `RetuningAutomations` to simplify the logic.
+    - Added 'LTE nodes with GUtranSyncSignalFrequency defined' to GUtran Frequency Audit
   
   - #### 🐛 Bug fixes:
     - Minor bug fixing.